+++ unrest/tests/flask/test_post.py
@@ -1,5 +1,5 @@
 from . import idsorted
-from ..model import Tree
+from ..model import Fruit, Tree
 
 
 def test_post_tree(rest, http):
@@ -144,3 +144,125 @@ def test_post_tree_custom_manual_commit(rest, http):
         'id': 3,
         'name': 'oak'
     }]
+
+
+def test_post_with_defaults(rest, http):
+    rest(
+        Fruit,
+        methods=['GET', 'POST'],
+        defaults={
+            'color': 'white',
+            'age': 42.0,
+        }
+    )
+    code, json = http.post('/api/fruit', json={'size': 1.0, 'tree_id': 1})
+    assert code == 200
+    assert json['occurences'] == 1
+    assert idsorted(json['objects'], 'fruit_id') == [
+        {
+            'fruit_id': 6,
+            'color': 'white',
+            'size': 1.0,
+            'age': 42.0,
+            'tree_id': 1
+        },
+    ]
+    code, json = http.post(
+        '/api/fruit', json={
+            'color': 'yellow',
+            'size': 2.0,
+            'tree_id': 2
+        }
+    )
+    assert code == 200
+    assert json['occurences'] == 1
+    assert idsorted(json['objects'], 'fruit_id') == [
+        {
+            'fruit_id': 7,
+            'color': 'yellow',
+            'size': 2.0,
+            'age': 42.0,
+            'tree_id': 2
+        },
+    ]
+
+    code, json = http.get('/api/fruit')
+    assert code == 200
+    assert json['occurences'] == 7
+    assert idsorted(json['objects'][-2:], 'fruit_id') == [
+        {
+            'fruit_id': 6,
+            'color': 'white',
+            'size': 1.0,
+            'age': 42.0,
+            'tree_id': 1
+        },
+        {
+            'fruit_id': 7,
+            'color': 'yellow',
+            'size': 2.0,
+            'age': 42.0,
+            'tree_id': 2
+        },
+    ]
+
+
+def test_post_with_fixed(rest, http):
+    rest(
+        Fruit,
+        methods=['GET', 'POST'],
+        fixed={
+            'color': 'white',
+            'age': 42.0,
+        }
+    )
+    code, json = http.post('/api/fruit', json={'size': 1.0, 'tree_id': 1})
+    assert code == 200
+    assert json['occurences'] == 1
+    assert idsorted(json['objects'], 'fruit_id') == [
+        {
+            'fruit_id': 6,
+            'color': 'white',
+            'size': 1.0,
+            'age': 42.0,
+            'tree_id': 1
+        },
+    ]
+    code, json = http.post(
+        '/api/fruit', json={
+            'color': 'yellow',
+            'size': 2.0,
+            'tree_id': 2
+        }
+    )
+    assert code == 200
+    assert json['occurences'] == 1
+    assert idsorted(json['objects'], 'fruit_id') == [
+        {
+            'fruit_id': 7,
+            'color': 'white',
+            'size': 2.0,
+            'age': 42.0,
+            'tree_id': 2
+        },
+    ]
+
+    code, json = http.get('/api/fruit')
+    assert code == 200
+    assert json['occurences'] == 7
+    assert idsorted(json['objects'][-2:], 'fruit_id') == [
+        {
+            'fruit_id': 6,
+            'color': 'white',
+            'size': 1.0,
+            'age': 42.0,
+            'tree_id': 1
+        },
+        {
+            'fruit_id': 7,
+            'color': 'white',
+            'size': 2.0,
+            'age': 42.0,
+            'tree_id': 2
+        },
+    ]