+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.7.2](https://github.com/Kozea/unrest/compare/0.7.1...0.7.2)
+
+* Add a defaults attribute to set column defaults.
+* Add a fixed attribute to force columns values.
+
 ## [0.7.1](https://github.com/Kozea/unrest/compare/0.7.0...0.7.1)
 
 * Add a manual commit flag to `@declare` to prevent auto commit.